+++ packages/stacked_generator/lib/src/generators/enums/dependency_type.dart
@@ -3,4 +3,5 @@ enum DependencyType {
   LazySingleton,
   Singleton,
   PresolvedSingleton,
+  FactoryWithParam,
 }